requirements.txt
streamlit
google-generativeai
openapi-core
PyYAML
jsonschema
pydantic
requests
toml